Week 5/Task 1/requirements.txt
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-openai
langchain-aws
langgraph
faiss-cpu
tavily-python
boto3
python-dotenv
numpy
tiktoken
